lavf/gifdec: do not mark as notimestamps
The demuxer does not set packet timestamps itself after
fc8e61645dfc4708c54f150597ce59ba7a2387b4 and instead relies on the
parser to do it. However, this does not matter from the caller
perspective as it still happens inside the demuxer. The demuxer should
thus not be flagged as not having timestamps.