lavc/h264dsp: fix R-V V weight_pixels pointer arithmetic
As of d88dc3cb318f6770dee697cc02efdb312735cff7,
the code is unrolled to process two rows per iteration.
The output cursor thus needs to be incremented by twice the
stride, which is taken care of with SH1ADD. However the original
ADD from the original implemetation was incorrectly left over.